requirements.txt
asgiref==3.11.0
Django==4.2
django-cors-headers==4.0.0
djangorestframework==3.14.0
numpy==2.0.2
pandas==2.3.3
psycopg2-binary==2.9.6
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.0
PyJWT==2.10.1
pytz==2025.2
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
djangorestframework-simplejwt==5.3.1
